+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/auth/AccountsRepository.kt
@@ -3,6 +3,6 @@ package com.sourcegraph.cody.auth
 /**
  * In most cases should be an instance of [com.intellij.openapi.components.PersistentStateComponent]
  */
-interface AccountsRepository<A: Account> {
+interface AccountsRepository<A : Account> {
   var accounts: Set<A>
 }