+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/config/notification/AccountSettingChangeActionNotifier.kt
@@ -3,14 +3,15 @@ package com.sourcegraph.cody.config.notification
 import com.intellij.util.messages.Topic
 
 interface AccountSettingChangeActionNotifier {
-    companion object {
-        @JvmStatic
-        val TOPIC = Topic.create(
+  companion object {
+    @JvmStatic
+    val TOPIC =
+        Topic.create(
             "Sourcegraph Cody + Code Search plugin settings have changed",
             AccountSettingChangeActionNotifier::class.java)
-    }
+  }
 
-    fun beforeAction(serverUrlChanged: Boolean)
+  fun beforeAction(serverUrlChanged: Boolean)
 
-    fun afterAction(context: AccountSettingChangeContext)
+  fun afterAction(context: AccountSettingChangeContext)
 }