+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/statusbar/CodyDisableAutocompleteAction.kt
@@ -7,13 +7,14 @@ import com.sourcegraph.cody.config.CodyApplicationSettings
 import com.sourcegraph.config.ConfigUtil
 
 class CodyDisableAutocompleteAction : DumbAwareAction() {
-    override fun actionPerformed(e: AnActionEvent) {
-        CodyApplicationSettings.getInstance().isCodyAutocompleteEnabled = false
-        CodyAutocompleteManager.getInstance().clearAutocompleteSuggestionsForAllProjects()
-    }
+  override fun actionPerformed(e: AnActionEvent) {
+    CodyApplicationSettings.getInstance().isCodyAutocompleteEnabled = false
+    CodyAutocompleteManager.getInstance().clearAutocompleteSuggestionsForAllProjects()
+  }
 
-    override fun update(e: AnActionEvent) {
-        super.update(e)
-        e.presentation.isEnabledAndVisible = ConfigUtil.isCodyEnabled() && ConfigUtil.isCodyAutocompleteEnabled()
-    }
+  override fun update(e: AnActionEvent) {
+    super.update(e)
+    e.presentation.isEnabledAndVisible =
+        ConfigUtil.isCodyEnabled() && ConfigUtil.isCodyAutocompleteEnabled()
+  }
 }