+++ cmd/frontend/internal/search/resolvers/search_job_stats.go
@@ -7,20 +7,24 @@ import (
 var _ graphqlbackend.SearchJobStatsResolver = &searchJobStatsResolver{}
 
 type searchJobStatsResolver struct {
+	total      int32
+	completed  int32
+	failed     int32
+	inProgress int32
 }
 
 func (e *searchJobStatsResolver) Total() int32 {
-	return 0
+	return e.total
 }
 
 func (e *searchJobStatsResolver) Completed() int32 {
-	return 0
+	return e.completed
 }
 
 func (e *searchJobStatsResolver) Failed() int32 {
-	return 0
+	return e.failed
 }
 
 func (e *searchJobStatsResolver) InProgress() int32 {
-	return 0
+	return e.inProgress
 }